client/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
alembic==1.14.1
python-dotenv==1.0.1
requests==2.32.3
PySocks==1.7.1
feedparser==6.0.11
dnslib==0.9.25
dnspython==2.6.1
git+https://github.com/ppouria/persian-encoder.git@aafcdad87bef8df598c66ebbd4e827feb664c71e
